requirements.txt
telethon
--only-binary=:all: pandas>=2.2.0
python-dotenv
regex>=2023.12.25

